Integrate ed25519-donna (Not yet used).
Integrate ed25519-donna into the build process, and provide an
interface that matches the `ref10` code.  Apart from the blinding and
Curve25519 key conversion, this functions as a drop-in replacement for
ref10 (verified by modifying crypto_ed25519.c).

Tests pass, and the benchmarks claim it is quite a bit faster, however
actually using the code more than likely requires sanity checking donna
vs ref10 at runtime.

NB: This probably breaks the Visual Studio build, but fixing that is
left as an exercise for someone who gives a fuck.