squash! Add blinding support to ed25519-donna (Not yet used).
Add `ALIGN(16)` directives in strategic locations so that the SSE2 code
will work.  It's worth noting that the SSE2 code can end up being
slower on certain systems than the portable variant (per the
documentation SSE2 is appears to be a win on 32 bit systems, but not 64
bit).

I'm undecided as to if requiring SSE2 is an ok thing to do for 32 bit
builds, since certain people will undoubtedly complain that tor no
longer works on their mid-90s shitbox.