Refactor loop over cpath for getting nicknames into separate function.
This refactors `circuit_list_path_impl()` to move the loop over the
`hops` in a `circ->cpath` to a separate function.  This is necessary
because retrieving router nicknames/fingerprints from a chosen cpath
requires the same process — regardless of what type of `circuit_t` the
`cpath` is an attribute of.  With these changes, other types of
`circuit_t`s — not solely `origin_circuit_t` — can reuse this giant code
block for logging their chosen paths.  In the future, this will be used
for logging `loose_or_circuit_t`s for prop#188.

 * ADDS a new function, `circuit_list_cpath_impl()`, which takes
   `crypt_path_t *cpath` and a `smartlist_t *elements`, iterates over
   the `cpath` to retrieve information regarding the hops within, and
   adds this information to the list of `elements`.

 * ADD a wrapper function `circuit_list_cpath()`, which wraps
   `circuit_list_cpath_impl()`.

 * REFACTORS `circuit_list_path_impl()` to call `circuit_list_cpath()`,
   giving it the former's pre-allocated `smartlist_t` in order to gather
   router info.  It then uses to build a string suitable for logging the
   circuit's path (just as it did before).