Add min and max weights to the fallback directory script
Modify fallback directory update script to:
* enforce a minimum fallback count
* reweight relays with high weights (avoid them seeing too many clients)
* exclude relays with low weights

If fallback weight criteria can't be satisfied, output #error if
STRICT_FALLBACK_WEIGHTS is True, and a comment if it is False.

Produce a summary comment at the top of the generated file,
describing how applying the minimum and maximum counts and weights
has affected the list of fallback directories.