Add fallback directory blacklist and whitelist
Entries in the whitelist are included if all attributes match
(with IPv6 addresses being optional).
Entries in the blacklist are excluded if any sufficiently specific
attribute group matches:
IPv4 & DirPort
IPv4 & ORPort
ID
IPv6 & DirPort
IPv6 & ORPort
If the ports are not present, the entire IP address is blacklisted.

We exclude entries unless explicitly listed in the whitelist, as we
are running an opt-in trial.

Also set trial parameters so fallbacks see between 10% and 0.01% of
requests.