fixup! Prop210: Refactor connection_get_* to produce lists and counts
Remove the duplicate code in the get* functions using template macros.

This adds an extra check to connection_get_by_global_id(), but doesn't
change the behaviour of tor, because its only caller also makes this
check.

Modify the get* functions that compare strings to use strcmp_opt.

Remove the duplicate code in the header, by using a template macro.
This also fixes an issue where free() was used instead of
smartlist_free().