prop250: Add memory and disk state in new files
This commit introduces two new files with their header.

"shared-random.c" contains basic function to initialize the state and allow
commit decoding for the disk state to be able to parse them from disk.

"shared-random-state.c" contains everything that has to do with the state
for both our memory and disk state. Lots of helper functions as well as a
mechanism to query the state in a synchronized way.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>