prop250: Act post consensus when we set the current one
We need to inform the SR subsystem when we set the current consensus because
this can be called when loaded from file or downloaded from other
authorities or computed.

Before that, we were only acting post consensus when we computed one which
was wrong because it lead to the SR subsystem use the wrong voting interval.

Signed-off-by: David Goulet <dgoulet@torproject.org>