Client intro & HS rend opportunistically upgrade to ntor
Clients look up the intro's ntor key from the relay's descriptor,
if it is available. Otherwise, they use the TAP key for the intro
point in the HS descriptor.

Hidden services look up the rend point's ntor key from the relay's
descriptor, if it is available. Otherwise, they use the TAP key for
the rend point in the INTRODUCE cell.

Add notes about why it's ok to replace the extend_info sent by the
other side with one from our consensus.

Add a stub function to support direct Single Onion rend connections.
(It currently hard-codes "not a single onion service".)