Client intro opportunistically upgrades to ntor
Clients look up the intro's ntor key from the relay's descriptor,
if it is available. Otherwise, they use the TAP key for the intro
point in the HS descriptor.

Add notes about why it's ok to replace the extend_info sent by the
hidden service with one from the client's consensus.