Fuzz Tor using the instructions in doc/HACKING/Fuzzing.md
Adds fuzz_http_header, a fuzzing harness for directory header parsing,
and test cases in src/test/fuzz, which can be run using 'make fuzz'.

Still needs a changes file, and a fix for the bug exposed by one of the
test cases. (Or, rather, the bug should be fixed in a separate branch.)