Remove IPv6 addresses from microdescriptors when they are unreachable
Defines MIN_METHOD_FOR_REMOVE_UNREACHABLE_IPV6_FROM_MD in exactly the wrong
way, by making microdescriptors depend on more than just the routerinfo.

The correct way to fix this issue is to define a flag like UnreachableIPv6,
have authorities vote for it if AuthDirHasIPv6Connectivity is set, and
then have clients ignore the a line if UnreachableIPv6 is present.

Bad code for issue 20916.