Add the EVENT_PRIVCOUNT_HSDIR_CACHE_STORED event
This event is emitted when a HS descriptor is uploaded to a v2 or v3 HSDir.
It contains the descriptor ID (v2) or blinded public key (v3), intro point
count (v2 only), times, sizes, and caching and validation information.

It does not contain the hidden service address, or any circuit information.

Part of privcount/privcount#336.