Fail attempts to connect to unreachable addresses
When a client is configured with ReachableAddresses, fail any attempts to
connect to an unreachable ORPort address. But allow attempts to connect to
bridge addresses, because they are sometimes placeholders for a pluggable
transport. (Configuring pluggable transports to only connect to certain
addresses is outside the scope of tor.)

There is no equivalent DirPort code, because clients no longer use the
DirPort.

This is a precaution against future bugs like #23493, which could have
led to IPv6-only v3 single onion services connecting directly to IPv4
addresses.